refactor(vocab,opensidian): import the auth client from $lib/platform/auth
Both flat-root apps aliased $platform/auth to a single-impl
src/lib/platform/auth/auth.ts, a seam-shaped name with no platform
branching behind it (the real two-impl seam is #platform/auth in
Whispering and Honeycrisp). Flatten the one-file directory to
src/lib/platform/auth.ts, import it as $lib/platform/auth (the
documented flat-root convention), and delete the svelte.config alias.
