+++ apps/opensidian/src/lib/opensidian.ts
@@ -25,7 +25,7 @@
 
 import { createNodeId } from '@epicenter/workspace';
 import { openOpensidianBrowser } from 'opensidian/browser';
-import { auth } from '$platform/auth';
+import { auth } from '$lib/platform/auth';
 import { createEditorState } from './state/editor-state.svelte';
 import { createFilesState } from './state/files-state.svelte';
 import { createPaletteSearchState } from './state/palette-search-state.svelte';
