+++ apps/opensidian/src/lib/platform/auth.ts
