+++ apps/opensidian/svelte.config.js
@@ -5,10 +5,6 @@ import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 const config = {
 	preprocess: vitePreprocess(),
 	kit: {
-		alias: {
-			// kit.alias is the source of truth for Vite and generated TypeScript config.
-			'$platform/auth': './src/lib/platform/auth/auth.ts',
-		},
 		adapter: staticAdapter({
 			fallback: 'index.html',
 		}),
