+++ apps/vocab/src/lib/platform/auth.ts
