+++ apps/vocab/src/lib/state/inference-connections.svelte.ts
@@ -14,7 +14,7 @@ import { API_ROUTES } from '@epicenter/constants/api-routes';
 import { APP_URLS } from '@epicenter/constants/vite';
 import { createPersistedState } from '@epicenter/svelte';
 import { VOCAB_MODEL } from '@epicenter/vocab';
-import { auth } from '$platform/auth';
+import { auth } from '$lib/platform/auth';
 
 export const inferenceConnections = createInferenceConnections({
 	storageKey: 'vocab',
