+++ apps/vocab/src/routes/auth/callback/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { Loading } from '@epicenter/ui/loading';
 	import { goto } from '$app/navigation';
-	import { auth } from '$platform/auth';
+	import { auth } from '$lib/platform/auth';
 
 	let errorMessage = $state<string | null>(null);
