+++ apps/vocab/src/routes/components/VocabSidebar.svelte
@@ -7,7 +7,7 @@
 	import MessageSquareTextIcon from '@lucide/svelte/icons/message-square-text';
 	import TrashIcon from '@lucide/svelte/icons/trash';
 	import { instanceSetting } from '$lib/instance';
-	import { auth } from '$platform/auth';
+	import { auth } from '$lib/platform/auth';
 	import { dictation } from '$lib/state/dictation.svelte';
 	import { vocab } from '$lib/vocab';
 	import TermsPanel from './TermsPanel.svelte';
