+++ apps/vocab/svelte.config.js
@@ -8,9 +8,6 @@ const config = {
 		adapter: staticAdapter({
 			fallback: 'index.html',
 		}),
-		alias: {
-			'$platform/auth': './src/lib/platform/auth/auth.ts',
-		},
 	},
 };
